Phase 6e: Consolidate unused hooks, extract event list loader, DRY refresh pattern
Phase 6e — the final extraction phase — consolidates duplicate code with
existing unused hooks, extracts event list management into a new hook, and
DRYs up the shared event metadata refresh pattern.

Steps completed:

1. Consolidate useCheesyArena hook (Step 1)
   - App.jsx had duplicate state declarations (cheesyArenaAvailable,
     useCheesyArena, cheesyTeamList) and a duplicate getCheesyStatus
     function that were never connected to the existing useCheesyArena hook
   - Imported useCheesyArenaStatus into App.jsx, replaced all duplicate
     state and the getCheesyStatus function body with hook calls
   - App.jsx getCheesyStatus now delegates to the hook's probe function

2. Consolidate useFTCOfflineMode hook (Step 2)
   - App.jsx had duplicate state (FTCOfflineAvailable, useFTCOffline,
     FTCKey, FTCServerURL, manualOfflineMode, ftcLeagues, ftcTypes) and
     duplicate functions (getFTCOfflineStatus, getFTCLeagues, requestFTCKey,
     checkFTCKey) plus a duplicate useEffect for FTC offline status polling
   - Imported useFTCOfflineMode into App.jsx, removed all duplicate state,
     functions, and effects
   - ~107 lines removed

3. Extract useEventListLoader hook (Step 3)
   - Created src/hooks/useEventListLoader.js (~478 lines) containing:
     - getEvents (~291 lines): main event list fetcher with FRC/FTC/TBA
       support, event enrichment, and timezone lookup
     - mergeTBAWithFIRSTEvents (~79 lines): merges TBA offseason events
     - fetchTBAEvents (~12 lines): fetches TBA offseason events
     - getDistricts (~21 lines): fetches district list for FRC mode
   - Moved color constants (paleYellow, paleBlue) and timezone data
     (cloneDeep of timeZones) into the hook module
   - Wired hook into App.jsx with deps object including updateFtcRegions
     callback for module-level ftcregions mutation
   - Removed unused moment import from App.jsx
   - ~403 lines removed from App.jsx

4. Extract shared refreshEventMetadata helper (Step 4 — simplified)
   - loadEvent and the useInterval auto-refresh block both duplicated the
     same messages + high scores refresh pattern (~15 lines each)
   - Extracted a shared refreshEventMetadata() function that handles:
     getSystemMessages, getEventMessages, getWorldStats, and the
     FTC/FRC high scores branching logic
   - Both loadEvent and useInterval now call refreshEventMetadata()
   - Full loadEvent hook extraction was deferred — the function has 40+
     dependencies making a hook wrapper unwieldy for marginal benefit

Results:
- App.jsx: 3,192 → 2,651 lines (541 lines removed, -17%)
- New file: src/hooks/useEventListLoader.js (478 lines)
- Zero build warnings
- All existing functionality preserved
