+++ src/App.jsx
@@ -28,7 +28,6 @@ import { useSettings } from "./contexts/SettingsContext";
 import { useEventData } from "./contexts/EventDataContext";
 import { EventStoreProvider } from "./contexts/EventStoreProvider";
 import _ from "lodash";
-import moment from "moment";
 import Developer from "./pages/Developer";
 import { eventNames, FTCEventNames } from "./data/eventNames";
 import { specialAwards, hallOfFame, FTCHallOfFame } from "./data/hallOfFame";
@@ -38,7 +37,6 @@ import { appUpdates } from "./data/appUpdates";
 import { useOnlineStatus } from "./contextProviders/OnlineContext";
 import { toast } from "react-toastify";
 import { trainingData } from "components/TrainingMatches";
-import { timeZones } from "data/timeZones";
 import {
   getConnectionsEventKey,
   fetchAllianceConnections,
@@ -53,6 +51,9 @@ import { useHotkeys } from "react-hotkeys-hook";
 import { useNotifications } from "hooks/useNotifications";
 import { useHighScores } from "hooks/useHighScores";
 import { useCommunityUpdates } from "hooks/useCommunityUpdates";
+import { useCheesyArenaStatus } from "hooks/useCheesyArena";
+import { useFTCOfflineMode } from "hooks/useFTCOfflineMode";
+import { useEventListLoader } from "hooks/useEventListLoader";
 
 export const TabStates = {
   NotReady: "notready",
@@ -88,8 +89,6 @@ const FTCSupportedYears = [
   // { label: "2022", value: "2022", program: "FTC" },
 ];
 
-const paleYellow = "#fdfaed";
-const paleBlue = "#effdff";
 const ftcBaseURL = "https://api.gatool.org/ftc/v2/";
 
 // Pages that should remember scroll position
@@ -170,8 +169,6 @@ function LayoutsWithNavbar({
 }
 
 const training = _.cloneDeep(trainingData);
-
-const timezones = _.cloneDeep(timeZones);
 
 var ftcregions = _.cloneDeep(ftcRegions);
 const regionLookup = [];
@@ -520,13 +517,6 @@ function App() {
 
   const [ftcMode, setFTCMode] = usePersistentState("setting:ftcMode", null);
 
-  const [ftcLeagues, setFTCLeagues] = usePersistentState(
-    "cache:ftcLeagues",
-    []
-  );
-
-  const [ftcTypes, setFTCTypes] = usePersistentState("cache:ftcTypes", []);
-
   // Controllers for table sort order at render time
   const [teamSort, setTeamSort] = useState("");
   const [rankSort, setRankSort] = useState("");
@@ -545,29 +535,28 @@ function App() {
     ? _.cloneDeep(FTCHallOfFame)
     : _.cloneDeep(hallOfFame);
 
-  // Cheesy Arena status
-  const [cheesyArenaAvailable, setCheesyArenaAvailable] = useState(false);
-  const [useCheesyArena, setUseCheesyArena] = usePersistentState(
-    "setting:useCheesyArena",
-    null
-  );
-  const [cheesyTeamList, setCheesyTeamList] = useState([]);
+  // Cheesy Arena status (hook owns state + status probe)
+  const {
+    cheesyArenaAvailable,
+    useCheesyArena, setUseCheesyArena,
+    cheesyTeamList, setCheesyTeamList,
+    getCheesyStatus: cheesyStatusProbe,
+  } = useCheesyArenaStatus();
 
-  // FTC Offline status
-  const [FTCOfflineAvailable, setFTCOfflineAvailable] = useState(false);
-  const [useFTCOffline, setUseFTCOffline] = usePersistentState(
-    "setting:useFTCOffline",
-    false
-  );
-  const [FTCKey, setFTCKey] = usePersistentState("setting:FTCKey", {});
-  const [FTCServerURL, setFTCServerURL] = usePersistentState(
-    "setting:FTCServerURL",
-    "http://10.0.100.5"
-  );
-  const [manualOfflineMode, setManualOfflineMode] = usePersistentState(
-    "setting:manualOfflineMode",
-    false
-  );
+  // FTC Offline mode (hook owns state + functions + effects)
+  const {
+    FTCOfflineAvailable,
+    useFTCOffline, setUseFTCOffline,
+    FTCKey, setFTCKey,
+    FTCServerURL, setFTCServerURL,
+    manualOfflineMode, setManualOfflineMode,
+    ftcLeagues,
+    ftcTypes, setFTCTypes,
+    getFTCOfflineStatus,
+    getFTCLeagues,
+    requestFTCKey,
+    checkFTCKey,
+  } = useFTCOfflineMode({ httpClient, ftcMode, selectedYear, isOnline });
 
   // High scores state and functions
   const {
@@ -672,72 +661,42 @@ function App() {
     useCheesyArena,
   });
 
+  // Event list management (getEvents, getDistricts)
+  const updateFtcRegions = (newRegions) => { ftcregions = newRegions; };
+  const { getEvents, getDistricts } = useEventListLoader({
+    httpClient,
+    selectedYear,
+    ftcMode,
+    useFTCOffline,
+    FTCServerURL,
+    FTCKey,
+    isOnline,
+    training,
+    supportedYears,
+    selectedEvent,
+    eventsLoading,
+    eventnames,
+    regionLookup,
+    setEvents,
+    setEventsLoading,
+    setFTCTypes,
+    setEventNamesCY,
+    setDistricts,
+    setSelectedEvent,
+    updateFtcRegions,
+  });
+
   /**
-   * Function to get the Cheesy Arena status by connecting to the Cheesy Arena API
+   * Wrapper around the hook's getCheesyStatus probe — adds loadSchedule orchestration
    */
   const getCheesyStatus = async (loadSchedule) => {
-    // See if you can connect to Cheesy Arena
-    console.log("Checking Cheesy Arena status...");
-    try {
-      var result = await fetch(
-        "http://10.0.100.5:8080/api/matches/qualification",
-        {
-          signal: AbortSignal.timeout(5000),
-        }
-      );
-      var data = result.status === 200;
-      // Set the IP address to the constant `ip`
-      if (data) {
-        console.log("Cheesy Arena is available.");
-        setCheesyArenaAvailable(true);
-        if (loadSchedule) {
-          eventStoreRef.current.getTeamList();
-          eventStoreRef.current.getSchedule();
-        }
-        return true;
-      } else {
-        console.log("Cheesy Arena is not available.");
-        setCheesyArenaAvailable(false);
-        return false;
-      }
-    } catch (error) {
-      console.log("Error fetching Cheesy Arena status:", error.message);
-      setCheesyArenaAvailable(false);
-      return false;
+    const available = await cheesyStatusProbe();
+    if (available && loadSchedule) {
+      eventStoreRef.current.getTeamList();
+      eventStoreRef.current.getSchedule();
     }
+    return available;
   };
-
-  // eslint-disable-next-line
-  const getFTCOfflineStatus = async () => {
-    // See if you can connect to FTC Local Server
-    // Requires that FTC Server URL be set.
-    console.log("Checking FTC Local Server status...");
-    try {
-      const result = await httpClient.getNoAuth(
-        `/api/v1/version/`,
-        FTCServerURL,
-        5000,
-        { Authorization: FTCKey?.key || "" }
-      );
-      if (result?.status === 200) {
-        console.log("FTC Local Server is available.");
-        return setFTCOfflineAvailable(true);
-      } else {
-        console.log("FTC Local Server is not available.");
-        return setFTCOfflineAvailable(false);
-      }
-    } catch (error) {
-      console.log("Error fetching FTC Local Server status:", error.message);
-      return setFTCOfflineAvailable(false);
-    }
-  };
-
-  // Check the status of FTC Local Server when the URL changes
-  useEffect(() => {
-    if (ftcMode?.value === "FTCLocal" && FTCServerURL) {
-      getFTCOfflineStatus();
-    }
-  }, [FTCServerURL, ftcMode?.value]); // eslint-disable-line react-hooks/exhaustive-deps
 
   const allianceConnectionsPrefetchSignature = useMemo(() => {
     if (!alliances?.alliances?.length) return "";
@@ -810,26 +769,6 @@ function App() {
 
   //functions to retrieve API data
 
-  /**
-   * Helper function to fetch all TBA offseason events
-   * @param {string} year Year
-   * @returns Array of TBA events
-   */
-  const fetchTBAEvents = async (year) => {
-    try {
-      console.log(`Fetching all TBA offseason events for ${year}...`);
-      const result = await httpClient.getNoAuth(`${year}/offseason/events/`);
-      if (result.status === 200) {
-        // @ts-ignore
-        const tbaEvents = await result.json();
-        return tbaEvents.events || [];
-      }
-      return [];
-    } catch (error) {
-      console.error("Error fetching TBA events:", error);
-      return [];
-    }
-  };
   /**
    * Remaps a string team identifier to its numeric team number (e.g., "TeamA" -> 9990)
    * @param {string} teamString The string team identifier
@@ -1001,6 +940,26 @@ function App() {
 
 
 
+  // Shared helper: refresh messages + high scores for the current event.
+  // Used by both loadEvent and the auto-refresh interval.
+  const refreshEventMetadata = () => {
+    getSystemMessages();
+    getEventMessages();
+    getWorldStats();
+    if (ftcMode) {
+      getFTCRegionHighScores();
+      if (selectedEvent?.value?.leagueCode) {
+        getFTCLeagueHighScores();
+      } else {
+        setFtcLeagueHighScores(null);
+      }
+    } else if (selectedEvent?.value?.districtCode) {
+      getFrcDistrictHighScores();
+    } else {
+      setFrcDistrictHighScores(null);
+    }
+  };
+
   /**
    * This function loads an event when a user selects an event from the menu. It will reset all event data, load the event details, team lists, team updates, refresh scores, ranks and world stats when appropriate.
    * @async
@@ -1110,487 +1069,7 @@ function App() {
 
       eventStoreRef.current.getTeamList();
       await eventStoreRef.current.getSchedule(true);
-      getSystemMessages();
-      await getEventMessages();
-      getWorldStats();
-      if (ftcMode) {
-        getFTCRegionHighScores();
-        if (selectedEvent?.value?.leagueCode) {
-          getFTCLeagueHighScores();
-        } else {
-          setFtcLeagueHighScores(null);
-        }
-      } else if (selectedEvent?.value?.districtCode) {
-        getFrcDistrictHighScores();
-      } else {
-        setFrcDistrictHighScores(null);
-      }
-    }
-  };
-
-  /**
-   * Merges TBA offseason events with FIRST events
-   * - Matches TBA events to FIRST events by event code (firstEventCode), then by name
-   * - Adds tbaEventKey to matched FIRST events
-   * - Adds unmatched TBA events as new OffSeason events to the list
-   * @param {Array} firstEvents Array of FIRST events
-   * @param {Array} tbaEvents Array of TBA events
-   * @param {string} year Year for event code generation
-   * @returns Merged array of events
-   */
-  const mergeTBAWithFIRSTEvents = (firstEvents, tbaEvents, year) => {
-    console.log(`Merging ${tbaEvents.length} TBA events with ${firstEvents.length} FIRST events...`);
-
-    // Clone FIRST events to avoid mutations
-    let mergedEvents = _.cloneDeep(firstEvents);
-    const unmatchedTBAEvents = [];
-
-    // Process each TBA event
-    tbaEvents.forEach((tbaEvent) => {
-      let matched = false;
-
-      // First, try to match by firstEventCode
-      const firstCodeMatch = _.findIndex(mergedEvents, (e) =>
-        e.code?.toLowerCase() === tbaEvent.firstEventCode?.toLowerCase()
-      );
-
-      if (firstCodeMatch >= 0) {
-        // Found a match by event code - add TBA event key
-        mergedEvents[firstCodeMatch].tbaEventKey = tbaEvent.code?.split(year)[1];
-        matched = true;
-      } else {
-        // Try to match by name
-        const nameMatch = _.findIndex(mergedEvents, (e) =>
-          e.name?.trim() === tbaEvent.name?.trim()
-        );
-
-        if (nameMatch >= 0) {
-          // Found a match by name - add TBA event key
-          mergedEvents[nameMatch].tbaEventKey = tbaEvent.code?.split(year)[1];
-          matched = true;
-        }
-      }
-
-      // If no match found, add as a new offseason event
-      if (!matched) {
-        unmatchedTBAEvents.push(tbaEvent);
-      }
-    });
-
-    // Add unmatched TBA events as new OffSeason events
-    unmatchedTBAEvents.forEach((tbaEvent) => {
-      // Create a FIRST-compatible event object from TBA event
-      const newEvent = {
-        eventId: null,
-        code: tbaEvent.code,
-        divisionCode: null,
-        name: tbaEvent.name,
-        remote: false,
-        hybrid: false,
-        fieldCount: null,
-        published: false,
-        type: "OffSeason",
-        typeName: "Offseason Event",
-        regionCode: null,
-        leagueCode: null,
-        districtCode: null,
-        venue: tbaEvent.venue || tbaEvent.location_name || null,
-        address: tbaEvent.address || null,
-        city: tbaEvent.city || null,
-        stateprov: tbaEvent.stateprov || tbaEvent.state_prov || null,
-        country: tbaEvent.country || null,
-        website: tbaEvent.website || null,
-        liveStreamUrl: null,
-        coordinates: null,
-        webcasts: tbaEvent.webcasts || null,
-        timezone: tbaEvent.timezone || null,
-        dateStart: tbaEvent.dateStart || tbaEvent.start_date ? moment(tbaEvent.dateStart || tbaEvent.start_date).format() : null,
-        dateEnd: tbaEvent.dateEnd || tbaEvent.end_date ? moment(tbaEvent.dateEnd || tbaEvent.end_date).format() : null,
-        tbaEventKey: tbaEvent.code?.split(year)[1],
-        weekNumber: null,
-        champLevel: "",
-      };
-
-      mergedEvents.push(newEvent);
-    });
-
-    console.log(`Merge complete. Total events: ${mergedEvents.length} (${unmatchedTBAEvents.length} new from TBA)`);
-    return mergedEvents;
-  };
-
-  const getEvents = async () => {
-    if (
-      eventsLoading === "" ||
-      eventsLoading !==
-      `${ftcMode ? ftcMode.value : "FRC"}-${selectedYear?.value}`
-    ) {
-      console.log(
-        `Loading ${ftcMode ? ftcMode.label : "FRC"} events list for for ${selectedYear?.value
-        }...`
-      );
-      setEventsLoading(
-        `${ftcMode ? ftcMode.label : "FRC"}-${selectedYear?.value}`
-      );
-      try {
-        let result;
-        if (useFTCOffline) {
-          const val = await httpClient.getNoAuth(
-            `/api/v1/events/`,
-            FTCServerURL
-          );
-          if (val.status === 200) {
-            // @ts-ignore
-            result = await val.json();
-          }
-          if (result?.eventCodes?.length > 0) {
-            let events = result.eventCodes.map(async (code) => {
-              const val = await httpClient.getNoAuth(
-                `/api/v1/events/${code}/`,
-                FTCServerURL,
-                undefined,
-                { Authorization: FTCKey?.key || "" }
-              );
-              if (val.status === 200) {
-                // @ts-ignore
-                const localEvent = await val.json();
-                return {
-                  eventId: localEvent.eventCode,
-                  code: localEvent.eventCode,
-                  divisionCode: null,
-                  name: localEvent.name,
-                  remote: false,
-                  hybrid: false,
-                  fieldCount: localEvent.fieldCount,
-                  published: false,
-                  type: localEvent.type,
-                  typeName: "Local Event",
-                  regionCode: null,
-                  leagueCode: null,
-                  districtCode: null,
-                  venue: null,
-                  address: null,
-                  city: null,
-                  stateprov: null,
-                  country: null,
-                  website: null,
-                  liveStreamUrl: null,
-                  coordinates: null,
-                  webcasts: null,
-                  timezone: null,
-                  dateStart: moment(localEvent.start).format(),
-                  dateEnd: moment(localEvent.end).format(),
-                };
-              }
-            });
-            await Promise.all(events).then((values) => {
-              result.events = values;
-            });
-          } else {
-            result = { events: [] };
-          }
-        } else {
-          const val = await httpClient.getNoAuth(
-            `${selectedYear?.value}/events`,
-            ftcMode ? ftcBaseURL : null
-          );
-          if (val.status === 200) {
-            // @ts-ignore
-            result = await val.json();
-          }
-        }
-
-        if (typeof result.Events !== "undefined") {
-          result.events = result.Events;
-          delete result.Events;
-        }
-        var timeNow = moment();
-
-        // Add training events for current season only
-        if (selectedYear?.value === supportedYears[0].value && !ftcMode) {
-          result.events = result?.events.concat(training.events.events);
-        } else if (!ftcMode) {
-          // Add OFFLINE event for all FRC seasons
-          const offlineEvent = training.events.events.find(e => e.code === "OFFLINE");
-          if (offlineEvent) {
-            result.events = result?.events.concat([offlineEvent]);
-          }
-        }
-
-        // Fetch and merge TBA offseason events for FRC mode
-        if (!ftcMode && isOnline) {
-          try {
-            const tbaEvents = await fetchTBAEvents(selectedYear?.value);
-            if (tbaEvents.length > 0) {
-              result.events = mergeTBAWithFIRSTEvents(result.events, tbaEvents, selectedYear?.value);
-            }
-          } catch (error) {
-            console.error("Error fetching/merging TBA events:", error);
-            // Continue with FIRST events only if TBA fetch fails
-          }
-        }
-
-        // FTC: kickoff events are not competition events; omit from picker and type filters
-        if (ftcMode) {
-          result.events = _.filter(result?.events, (e) => {
-            const type = String(
-              _.get(e, "type") ?? _.get(e, "Type") ?? ""
-            )
-              .trim()
-              .toUpperCase();
-            const typeName = String(
-              _.get(e, "typeName") ?? _.get(e, "TypeName") ?? ""
-            )
-              .trim()
-              .toUpperCase();
-            return type !== "12" && typeName !== "KICKOFF";
-          });
-        }
-
-        var regionCodes = [];
-        var types = [];
-        const events = result?.events.map((e) => {
-          var color = "";
-          /** @type {null | 'offseasonAzure' | 'offseason'} */
-          var eventMenuTint = null;
-          var optionPrefix = "";
-          var optionPostfix = "";
-          var filters = [];
-
-          // We have four formats available in timezones: abbreviation, description, Livemeeting and Windows. We lookup the Windows
-          // format and convert it to a more standard format. Consider moving off of Moment on to Luxor?
-
-          e.timeZoneAbbreviation =
-            timezones[
-              _.findIndex(timezones, { Windows: e.timezone })
-            ]?.Abbreviation;
-
-          var eventTime = e.dateEnd ? moment(e.dateEnd) : moment();
-          e.name = e.name.trim();
-          e.name = _.replace(e.name, `- FIRST Robotics Competition -`, `-`);
-          e.name = _.replace(
-            e.name,
-            `FIRST Championship - FIRST Robotics Competition`,
-            `FIRST Championship - Einstein`
-          );
-          if (e.code === "week0" || e.code === "WEEK0") {
-            filters.push("week0");
-          }
-          if (e.type === "OffSeasonWithAzureSync") {
-            color = paleBlue;
-            eventMenuTint = "offseasonAzure";
-            optionPrefix = "•• ";
-            optionPostfix = " ••";
-            filters.push("offseason");
-          }
-          if (e.type === "OffSeason" || e.type === "10") {
-            color = paleYellow;
-            eventMenuTint = "offseason";
-            optionPrefix = "•• ";
-            optionPostfix = " ••";
-            filters.push("offseason");
-          }
-          if (e.type === "Regional") {
-            filters.push("regional");
-          } else if (e.type.startsWith("Champion")) {
-            filters.push("champs");
-          } else if (e.districtCode) {
-            filters.push("district");
-            filters.push(e.districtCode);
-          } else if (ftcMode) {
-            filters.push(e.type);
-            filters.push(e.leagueCode);
-            filters.push(e.regionCode);
-            regionCodes.push({
-              regionCode: e.regionCode,
-              description: regionLookup[e.regionCode]
-                ? `${regionLookup[e.regionCode]} (${e.regionCode})`
-                : e.regionCode,
-            });
-            types.push({ type: e.type, description: e.typeName });
-          }
-
-          if (timeNow.diff(eventTime) < 0) {
-            filters.push("future");
-          } else {
-            filters.push("past");
-          }
-          if (
-            eventTime.diff(timeNow, "days") <= 7 &&
-            eventTime.diff(timeNow, "days") >= -0
-          ) {
-            filters.push("thisWeek");
-          }
-          if (
-            eventTime.diff(timeNow, "weeks") <= 4 &&
-            eventTime.diff(timeNow, "weeks") >= 0
-          ) {
-            filters.push("thisMonth");
-          }
-          if (e.type !== "OffSeason" && e.type !== "OffSeasonWithAzureSync") {
-            filters.push("week" + e.weekNumber);
-          }
-
-          e.champLevel = "";
-
-          // new method using event type data
-          if (
-            e.type === "DistrictChampionship" ||
-            e.type === "DistrictChampionshipWithLevels"
-          ) {
-            e.champLevel = "DISTCHAMPS";
-          } else if (e.type === "DistrictChampionshipDivision") {
-            e.champLevel = "DISTDIV";
-          } else if (e.type === "ChampionshipDivision") {
-            e.champLevel = "CMPDIV";
-          } else if (e.type === "ChampionshipSubdivision") {
-            e.champLevel = "CMPSUB";
-          } else if (e.type === "Championship") {
-            e.champLevel = "CHAMPS";
-          } else if (e.type === "6") {
-            e.champLevel = "CHAMPS";
-          }
-
-          return {
-            value: e,
-            label: `${optionPrefix}${e.name}${optionPostfix}`,
-            color: color,
-            eventMenuTint: eventMenuTint,
-            filters: filters,
-          };
-        });
-
-        // use for diagnostics to find missing regionCodes
-        // regionCodes = _.filter(_.uniqBy(regionCodes,"regionCode"),function(o) {return _.filter(ftcregions, {regionCode:o.regionCode}).length === 0;});
-
-        ftcregions = _.orderBy(
-          _.uniqBy(regionCodes, "regionCode"),
-          "description",
-          "asc"
-        );
-        types = _.orderBy(_.uniqBy(types, "type"), "description", "asc");
-        setFTCTypes(types);
-
-        //Ensure that current year event names change when Division or sponsor names change
-        if (typeof eventnames[selectedYear?.value] === "undefined") {
-          eventnames[selectedYear?.value] = {};
-        }
-
-        events.forEach((event) => {
-          // if (!eventnames[selectedYear?.value]) {
-          //   eventnames[selectedYear?.value] = {};
-          // }
-
-          eventnames[selectedYear?.value][event?.value?.code] =
-            event?.value?.name;
-        });
-        setEventNamesCY(eventnames[selectedYear?.value]);
-
-        setEvents(events);
-        // Check if the currently selected event still exists in the newly loaded events list
-        if (selectedEvent) {
-          const matchingEvent = _.find(events, (event) => event?.value?.code === selectedEvent?.value?.code);
-          if (matchingEvent) {
-            // Update the selected event with fresh data from the API
-            console.log("Updating selected event with fresh data from API");
-            setSelectedEvent(matchingEvent);
-          } else {
-            // Event no longer exists in the list, clear selection
-            console.log("Previously selected event not found in current events list, clearing selection");
-            setSelectedEvent(null);
-            setEventsLoading("");
-          }
-        }
-      } catch (e) {
-        console.error(e);
-      }
-    } else {
-      console.log(
-        `Events already loaded for ${ftcMode ? ftcMode.label : "FRC"}-${selectedYear?.value
-        }. Skipping...`
-      );
-    }
-  };
-
-  const getDistricts = async () => {
-    try {
-      const val = await httpClient.getNoAuth(
-        `${selectedYear?.value}/districts`
-      );
-      if (val.status === 200) {
-        // @ts-ignore
-        const json = await val.json();
-        if (typeof json.Districts !== "undefined") {
-          json.districts = json.Districts;
-          delete json.Districts;
-        }
-        const districts = json.districts.map((district) => {
-          return { label: district.name, value: district.code };
-        });
-
-        setDistricts(districts);
-      }
-    } catch (e) {
-      console.error(e);
-    }
-  };
-
-  const getFTCLeagues = async () => {
-    // Skip external API calls when in FTC offline mode without internet
-    if (useFTCOffline && (!isOnline || manualOfflineMode)) {
-      console.log("FTC Offline mode: Skipping FTC Leagues API call while offline" + (manualOfflineMode ? " (manual override)" : "") + " - using cached leagues");
-      return;
-    }
-
-    try {
-      const val = await httpClient.getNoAuth(
-        `${selectedYear?.value}/leagues`,
-        ftcMode ? ftcBaseURL : undefined
-      );
-      if (val.status === 200) {
-        // @ts-ignore
-        const json = await val.json();
-        if (typeof json.Leagues !== "undefined") {
-          json.leagues = json.Leagues;
-          delete json.Leagues;
-        }
-        const leagues = json.leagues.map((league) => {
-          return { label: league.name, value: league.code };
-        });
-
-        setFTCLeagues(leagues);
-      }
-    } catch (e) {
-      console.error(e);
-    }
-  };
-
-  // Request FTC Key from local FTC Server
-  const requestFTCKey = async () => {
-    const val = await httpClient.postNoAuth(
-      `/api/v1/keyrequest/?name=gatool`,
-      null,
-      FTCServerURL,
-      null
-    );
-    if (val.status === 200) {
-      // @ts-ignore
-      const json = await val.json();
-      setFTCKey({ ...json, FTCServerURL: FTCServerURL, active: false });
-    }
-  };
-
-  // Get the FTC Key from gatool Cloud
-  const checkFTCKey = async () => {
-    const val = await httpClient.getNoAuth(
-      `/api/v1/keycheck/`,
-      FTCServerURL,
-      undefined,
-      { Authorization: FTCKey?.key || "" }
-    );
-    if (val.status === 200) {
-      // @ts-ignore
-      const json = await val.json();
-      setFTCKey({ ...FTCKey, active: json?.active });
+      refreshEventMetadata();
     }
   };
 
@@ -1697,14 +1176,6 @@ function App() {
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [ftcMode, selectedYear, httpClient]);
-
-  // Reset manual offline mode when switching to FRC or FTC Online mode
-  useEffect(() => {
-    if (ftcMode?.value !== "FTCLocal" && manualOfflineMode) {
-      console.log("Switching to online mode, resetting manual offline mode");
-      setManualOfflineMode(false);
-    }
-  }, [ftcMode, manualOfflineMode, setManualOfflineMode]);
 
   // Load FTC avatars composed CSS when in ftcMode (season+1 for next-season avatar styles)
   useEffect(() => {
@@ -1980,19 +1451,7 @@ function App() {
       } else {
         console.log("Offline event. Just get the world stats if you can");
       }
-      getSystemMessages();
-      getEventMessages();
-      getWorldStats();
-      if (ftcMode) {
-        getFTCRegionHighScores();
-        if (selectedEvent?.value?.leagueCode) {
-          getFTCLeagueHighScores();
-        } else {
-          setFtcLeagueHighScores(null);
-        }
-      } else if (selectedEvent?.value?.districtCode) {
-        getFrcDistrictHighScores();
-      }
+      refreshEventMetadata();
     },
     intervalDelayMs,
     {
